refactor : 불필요한 코드 수정
main.swift
refactor :
1. start 변수를 사용할 필요가 없어 삭제함
2. while에 탈출 조건을 작성하여 더욱 간결한 코드를 작성함

Utility.swift
refactor :
1. RPS.swift 내 isRepeat()을 해당 파일로 이동함
2. MukJjiBba.swift 내 shouldByClose()를 해당 파일로 이동함
